Rename helpers module to `prng`
The only method left is a pseudorandom number generator, which will
remain as starter code given to students.
