+++ code/part-two/processor/services/prng.js
@@ -16,6 +16,4 @@ const getPrng = hex => {
   };
 };
 
-module.exports = {
-  getPrng
-};
+module.exports = getPrng;
